move DreamGeometry and Test to separate files
